+++ packages/elasticsearch/src/example-types/metricGroups/stats.js
@@ -1,10 +1,10 @@
-let { statsAggs, simplifyAggregations } = require('../../utils/elasticDSL')
+let { statsAggs, simplifyBucket } = require('../../utils/elasticDSL')
 
 let buildQuery = ({ statsField, stats }) => statsAggs(statsField, stats)
 
 let result = async (node, search) => {
   let response = await search(buildQuery(node))
-  return simplifyAggregations(response.aggregations)
+  return simplifyBucket(response.aggregations)
 }
 
 module.exports = {