+++ packages/elasticsearch/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.7.0
+* Fix bug where 0 and null value metrics wouldn't be flattened by dsl utils
+* Factor out groupStats util
+* Add pivot type
+
 # 1.6.3
 * Fixes a bug extracting es types from mappings across different elasticsearch versions.